+++ metadata.rb
@@ -5,15 +5,17 @@ maintainer_email 'cookbooks@rightscale.com'
 license          'Apache 2.0'
 description      'Installs/Configures a PHP application server'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '2.0.1'
+version          '2.0.2'
 source_url       'https://github.com/rightscale-cookbooks/rs-application_php' if respond_to?(:source_url)
 issues_url       'https://github.com/rightscale-cookbooks/rs-application_php/issues' if respond_to?(:issues_url)
 chef_version '>= 12.0' if respond_to?(:chef_version)
 
-depends 'marker'
+supports 'ubuntu', '>= 14.04'
+
+depends 'apache2', '~> 3.2.2'
 depends 'application'
 depends 'application_php', '~> 2.0.2'
-depends 'mysql2_chef_gem'
+depends 'mysql2_chef_gem', '>= 2.0.0'
 depends 'database'
 depends 'git'
 depends 'php'
@@ -26,6 +28,7 @@ depends 'curl'
 depends 'yum-mysql-community'
 depends 'yum-epel'
 depends 'yum-ius'
+depends 'apt'
 
 recipe 'rs-application_php::default', 'Installs/configures PHP application server'
 recipe 'rs-application_php::tags', 'Sets up application server tags used in a 3-tier deployment setup'